FlaskSample/requirements.txt
adal==1.1.0
asn1crypto==0.24.0
azure-common==1.1.16
azure-mgmt-storage==3.0.0
azure-nspkg==3.0.2
azure-storage-blob==1.3.1
azure-storage-common==1.3.0
azure-storage-nspkg==3.0.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
Flask==1.0.2
future==0.16.0
idna==2.7
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.8.0
MarkupSafe==1.0
msrest==0.6.1
msrestazure==0.5.0
oauthlib==2.1.0
Pillow==5.3.0
pip==10.0.1
pycparser==2.19
PyJWT==1.6.4
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.7.3
requests==2.19.1
requests-oauthlib==1.0.0
setuptools==39.0.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
